reorganize src files for assignment
